Remove use of Reader as passed to Read.
Instead, just hand it a byte slice and let it go hog wild with it since
it's going to read it in as one anyway.

This also fixes an error when reading >256 files at a time, which is
quite speedy but also results in errors because I forgot to close files
and I assume the GC wasn't finalizing them fast enough.